# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "ISA TERM PROJECT"
  },
  "host": "seanwallace.ca",
  "basePath": "/v1",
  "tags": [
    {"name": "posts",
      "description": "post Operations"
    },
    {
      "name": "comments",
      "description": "Comment Operations"
    },
    {
      "name": "user",
      "description": "Operations about user"
    },
    {
      "name": "admin",
      "description": "Info about api usage"
    }
  ],
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
    "/posts":{
      "get":{
        "tags": [
          "posts"
        ],
        "summary": "Get all posts",
        "description": "Returns all posts",
        "operationId": "getposts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"array",
              "items": {"$ref": "#/definitions/post"}
            }
          },
          "500":{
            "description":"Internal database error"
          }
        }
      },
      "post": {
        "tags": ["posts"],
        "summary": "Create a new post",
        "description": "",
        "operationId": "addpost",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "post Object that needs to be added to the site",
            "required": true,
            "schema": {
              "$ref": "#/definitions/newPost"
            }
          }
        ],
        "responses": {
          "201":{
            "description": "New post created"
          },
          "405":{
            "description" :"Invalid Input"
          },
          "500":{
            "description": "Internal Database error"
          }
        }
      },
      "put": {
        "tags": ["posts"],
        "summary": "Update a post",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "post Object that needs to be updated on the site",
            "required": true,
            "schema": {
              "$ref": "#/definitions/post"
            }
          }
        ],
        "responses": {
          "200":{
            "description": "Successfull operation"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "post not found"
          },
          "405": {
            "description": "Validation exception"
          },
          "500":{
            "description": "Internal database error"
          }
        }
      }
    },
    "/posts/{postID}":{
      "delete": {
        "tags": ["posts"],
        "summary": "Delete a post",
        "description": "",
        "operationId": "deletepost",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "postID",
            "in": "path",
            "description": "post id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200":{
            "description": "post successfully deleted"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "post not found"
          }
      }
    }
    },
    "/comments":{
      "post": {
        "tags": ["comments"],
        "summary": "Create a new comment",
        "description": "",
        "operationId": "addComment",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Comment object too be added to a post",
            "required": true,
            "schema": {
              "$ref": "#/definitions/comment"
            }
          }
        ],
        "responses": {
          "201":{
            "description": "New comment created"
          },
          "405":{
            "description" :"Invalid Input"
          }
        }
      },
      "put": {
        "tags": ["comments"],
        "summary": "Update a comment",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Comment Object that needs to be updated on the site",
            "required": true,
            "schema": {
              "$ref": "#/definitions/comment"
            }
          }
        ],
        "responses": {
          "200":{
            "description": "Successfully updated the comment"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Comment not found"
          },
          "405": {
            "description": "Validation exception"
          }
        }
      }
    },
    "/comments/commentsForPost/{postID}":{
      "get":{
        "tags": [
          "comments"
        ],
        "summary": "Get comments for a post",
        "description": "Return the comments for a post",
        "operationId": "getCommentsBypost",
        "produces": [
          "application/json"
        ],
        "parameters": [{
          "name": "postID",
          "in": "path",
          "description": "post id to get comment",
          "required": true,
          "type": "integer",
          "format": "int64"
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type":"array",
              "items": {"$ref": "#/definitions/comment"}
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "post not found"
          },
          "500":{
            "description": "Internal Database error"
          }
        }
      }
    },
    "/comments/{commentID}":{
      "delete": {
        "tags": ["comments"],
        "summary": "Delete a comment",
        "description": "",
        "operationId": "deleteComment",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "commentID",
            "in": "path",
            "description": "comment id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200":{
            "description": "Comment successfully deleted"
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Comment not found"
          },
          "500":{
            "description": "Internal database error"
          }
      }
    }
    },
    "/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Create user",
        "description": "This can only be done by the logged in user.",
        "operationId": "createUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/NewUser"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "New user created"
          },
          "400":{
            "description": "Error message"
          },
          "500":{
            "description": "Internal database error"
          }
        }
      },
      "put": {
        "tags": [
          "user"
        ],
        "summary": "Updated user",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Updated user object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200":{
            "description": "Successfully updated user"
          },
          "400": {
            "description": "Invalid user supplied"
          },
          "404": {
            "description": "User not found"
          },
          "500":{
            "description": "Internal Database error"
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into the system",
        "description": "",
        "operationId": "loginUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name":"body",
            "description": "The login info",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            }

          },
          "400": {
            "description": "Invalid username/password supplied"
          },
          "500":{
            "description": "Internal database error"
          }
        }
      }
    },
    "/user/logout": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs out current logged in user session",
        "description": "",
        "operationId": "logoutUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/user/{username}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by user name",
        "description": "",
        "operationId": "getUserByName",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/User"
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          },
          "500":{
            "description": "Internal database error"
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "summary": "Delete user",
        "description": "This can only be done by the logged in user.",
        "operationId": "deleteUser",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "description": "The name that needs to be deleted",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          },
          "500":{
            "description": "Internal database error"
          }
        }
      }
    },
    "/admin":{
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get api usage",
        "description": "",
        "operationId": "getApi",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/admin"
            }
          },
          "500": {
            "description": "Database error"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
    "post":{
      "type": "object",
      "properties": {
        "id":{
          "type":"integer",
          "format": "int64"
        },
        "title":{
          "type":"string"
        },
        "content":{
          "type":"string"
        },
        "user":{
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "newPost":{
      "type": "object",
      "properties": {
        "title":{
          "type":"string"
        },
        "content":{
          "type":"string"
        },
        "user":{
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "comment":{
      "type": "object",
      "properties": {
        "postID":{
          "type": "string"
        },
        "id":{
          "type":"integer",
          "format": "int64"
        },
        "content":{
          "type":"string"
        },
        "user":{
          "type":"string"
        }
        
      }
    },
    "NewUser": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "repeatPassword":{
          "type":"string"
        },
        "authorization":{
          "type": "string",
          "enum": [
            "admin",
            "user"
          ]
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "username": {
          "type": "string"
        },
        "authorization":{
          "type": "string",
          "enum": [
            "admin",
            "user"
          ]
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "login": {
      "type": "object",
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "xml": {
        "name": "User"
      }
    },
    "admin":{
      "type":"object",
      "properties": {
        "get/posts": {
          "type": "integer",
          "format": "int32"
        },
        "post/posts": {
          "type": "integer",
          "format": "int32"
        },
        "put/posts": {
          "type": "integer",
          "format": "int32"
        },
        "delete/posts/{postID}": {
          "type": "integer",
          "format": "int32"
        },
        "post/comments": {
          "type": "integer",
          "format": "int32"
        },
        "put/comments": {
          "type": "integer",
          "format": "int32"
        },
        "get/comments/commentsForPost/{postID}": {
          "type": "integer",
          "format": "int32"
        },
        "delete/comments/{commentID}}": {
          "type": "integer",
          "format": "int32"
        },
        "post/user": {
          "type": "integer",
          "format": "int32"
        },
        "post/user/login": {
          "type": "integer",
          "format": "int32"
        },
        "post/user/logout": {
          "type": "integer",
          "format": "int32"
        },
        "get/user/{username}": {
          "type": "integer",
          "format": "int32"
        },
        "put/user/{username}": {
          "type": "integer",
          "format": "int32"
        },
        "delete/user/{username}": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
  }